requirements.txt
PyPDF2==3.0.1
langchain==0.1.20
sentence-transformers==2.6.1
faiss-cpu==1.7.4
transformers==4.41.2
numpy==1.26.4
streamlit==1.35.0
